old-tests/submission/W3C/domhtml/HTMLDocument02.html

<!DOCTYPE html>
<html><head>
<meta charset="utf-8">
<title>NIST DOM HTML Test - DOCUMENT</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../common/domtests-common.js"></script>
<link href="http://www.w3.org/TR/html5/dom.html#dom-document-referrer" rel="help">
<link rel="author" title="NIST">
</head>
<body id="TEST-BODY"> 
<iframe id='iframe' src="/support/html/sample.html">
</iframe>

<script>
var domtest = async_test("The referrer attribute returns the URI of the page that linked to this page.");

on_event(window, "load", domtest.step_func(function (dom_event) {
    var nodeList;
      var testNode;
      var vreferrer;
      var doc;
      doc = window.document.getElementById("iframe").contentDocument;
      assert_not_equals(doc,null);
      vreferrer = doc.referrer;

      assert_uri_equals(document.location.protocol,null,document.location.host,"HTMLDocument02.html",null,null,null,true,vreferrer);
    domtest.done();
}));

</script>
<p>The referrer attribute returns the URI of the page that linked to this
    page.

    Retrieve the referrer attribute and examine its value.</p>
<div id="log"></div>
</body></html>
